requirements.txt
certifi==2020.6.20
cycler==0.10.0
graphviz==0.15
install==1.3.4
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
neat-python==0.92
numpy==1.19.2
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
